Add xml:ids to resources
Add xml:ids to resources. This is a new requirement that has to do
with the generation of PDFs from WADL files – without this, the paths
are not unique across WADL files

Remove files v1/src/samples/os-volume-1.wadl and
v2/src/samples/os-volume-1.wadl which are not needed in the samples
directory - instead of adding xml:ids to them.

Change-Id: Idf0cd1aef4c7c65c2f09688a8163bb08d3dbf37f
Related-Bug: #1275007